requirements.txt
matplotlib==3.7.1
pandas==2.1.4
pytorch_minimize==0.0.2
torch==2.0.1+cu118
